feat: ButtonViewState.kt is moved to state package
